requirements.txt
json-five
cobs
pyserial
protobuf==3.20.3
